API: use a linker script to compile the demo
When compiling the API demo program, the first object file in the linker
arguments is crt0.o, which contains the _start routine. This is done
with the hope that it will appear first in the .text section, but the
linker doesn't guarantee this behaviour.

Add a simple linker script that ensures this ordering. This fixes
execution of the API demo binary for cases where _start was present
elsewhere in the .text section.

Signed-off-by: Mitchell Horne <mhorne@FreeBSD.org>